refactor: web app clean up
