+++ other/train-to-cloud-city/app/client/src/components/Main.jsx
@@ -26,14 +26,14 @@ import {
   updateInputMailbox,
   proposalUpdated,
 } from "../actions/coreActions";
-import ConductorWave from "../assets/conductor-wave.gif";
+import ConductorWave from "../assets/conductor-wave.svg";
 import "./styles/Main.css";
 
 /**
  * Main
  * -----------------
  * Sets environment and data to populate into dashboard
- * which is in 3 different states (simulator, realtime, virtual input)
+ * which is in 3 different states (simulator, realtime from physical train)
  */
 const Main = (props) => {
   const state = useSelector((state) => state);
@@ -45,29 +45,20 @@ const Main = (props) => {
   const [signals, setSignals] = useState({});
   const [cargo, setCargo] = useState({});
   const [train, setTrain] = useState({});
-  const [pattern, setPattern] = useState({});
   const [proposal, setProposal] = useState({});
   const [trainMailbox, setTrainMailbox] = useState({});
 
-  const handleSimulator = async (event) => {
-    setToggle(!simulator);
-    setSimulator(!simulator);
-    dispatch?.(getWorldState(simulator ? "global_simulation" : "global"));
-  };
-
-  const cleanSlate = async () => {
-    try {
-      await updateInputMailbox("reset");
-      Promise.all([dispatch(getServices()), dispatch(getPatterns())]);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
   useEffect(() => {
+    const cleanSlate = async() => {
+      try {
+        await updateInputMailbox("reset");
+        Promise.all([dispatch(getServices()), dispatch(getPatterns())]);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+    
     const collection = simulator ? "global_simulation" : "global";
-
-    cleanSlate();
     // Listen for when patterns are updated
     proposalUpdated((data) => {
       setToggle(!!data.pattern_slug);
@@ -78,11 +69,18 @@ const Main = (props) => {
         cleanSlate();
       }
     }, collection);
+    
     signalsUpdated((data) => setSignals(data), collection);
     trainUpdated((data) => setTrain(data), collection);
     cargoUpdated((data) => setCargo(data), collection);
     trainMailboxUpdated((data) => setTrainMailbox(data), collection);
-  }, [simulator]);
+  }, [simulator, dispatch]);
+
+  const handleSimulator = async (event) => {
+    setToggle(!simulator);
+    setSimulator(!simulator);
+    dispatch?.(getWorldState(simulator ? "global_simulation" : "global"));
+  };
 
   return (
     <div className="mainContainer">
@@ -112,16 +110,13 @@ const Main = (props) => {
         {toggled ? (
           <Dashboard
             proposal={proposal}
-            pattern={pattern}
             train={train}
             cargo={cargo}
             signals={signals}
             trainMailbox={trainMailbox}
           />
         ) : (
-          <a href="#" onClick={handleSimulator}>
-            {"Turn on simulator"}
-          </a>
+          <a href="#" onClick={handleSimulator}>{"Turn on simulator"}</a>
         )}
       </div>
     </div>
