+++ other/train-to-cloud-city/app/client/src/components/Signal.jsx
@@ -16,7 +16,6 @@ import React from "react";
 import Station from "../assets/station.png";
 import SignalGreen from "../assets/signal-green.svg";
 import SignalOff from "../assets/signal-off.svg";
-import SignalYellow from "../assets/signal-yellow.svg";
 import SignalRed from "../assets/signal-red.svg";
 import "./styles/Signal.css";
 
@@ -27,7 +26,7 @@ import "./styles/Signal.css";
  */
 const Signal = (props) => {
   const { isStation, trainLocation, signal, showTrainLocation = true } = props;
-  const { name, slug, actual_state, target_state } = signal || {};
+  const { slug, target_state } = signal || {};
   const position = isStation ? "station" : slug;
   const checkpointClasses = slug === "four" ? "circle" : "circle connector";
   const locationClasses =
