+++ other/train-to-cloud-city/app/client/src/components/styles/Main.css
@@ -34,7 +34,6 @@
 }
 
 .welcomeImage img {
-  width: 250px;
   margin: 25px;
 }
